organize module tree
